requirements.txt
django==1.10.3
MySQL-python==1.2.5
djangorestframework==3.5.3
markdown==2.6.7
boto3==1.4.1
enum34==1.1.6
requests==2.12.1
pytz==2016.7
django-slack==5.7.1
Sphinx==1.4.8
sphinxcontrib-httpdomain==1.5.0
SQLAlchemy==1.1.4
marshmallow-sqlalchemy==0.12.0
django-redis-sessions-fork==0.7.0
drf-nested-routers==0.11.1
marshmallow